Fix a undefined access error in `finishGame`
This change gracefully exits with an error when finishGame is called on an invalid game, rather than creating an internal server error.

This error is actually fairly common and the only non-disconnect error in our functions right now. It occurs 96K times per month. I'm not sure why this is happening, but the first step is always to provide a better error message and not to crash.