Update game.ts
Only changed this file's functions. All the other files have to be adjusted yet.

I haven't been able to find a neater way for finding a Hyper set as done in ll. 94-99. Unfortunately there is no way like for the UltraSet, because three cards of the six Hyperset-cards do not have anything in common with the other three cards that you could feed a dictionary with...