Enable hints on Set-Chain and UltraSet (#178)
This enables hints on all game types. I actually don't remember why this wasn't enabled from the start. Anyway because this will disrupt about 20,000 games played in the profile pages that have enableHint set to `true` but were in set-chain/ultraset, I also wrote a migration script to fix this.